Clean up data by dropping variables systematically

This simple helper function will drop duplicated variable names created via data.table merges

drop_extramergevars(dt, string_match = "i\\.")

Arguments

dt

data.table or data.frame object

string_match

a string match to identify the set of additional variables to be dropped. "i\." will drop all variables that contain duplicated variables from a data.table merge