Another pulling function for pulling Google Earth Engine data for larger maps

Another pulling function for pulling Google Earth Engine data for larger maps

Usage

gee_pullbigdata(
  email = "ifeanyi.edochie@gmail.com",
  shp_dt,
  gee_name = "NOAA/VIIRS/DNB/MONTHLY_V1/VCMCFG",
  gee_datestart = "2019-01-01",
  gee_dateend = "2019-12-31",
  gee_band = "avg_rad",
  gee_chunksize = 4000,
  gee_stat = "mean",
  gee_scale = NULL
)

Arguments

email

A GEE authenticated gmail address

shp_dt

an sf/dataframe object

gee_datestart

The starting date for the specific feature of interest

gee_dateend

The ending date for the specific feature of interest

gee_chunksize

an integer; the size of each chunk which the shapefile should be divided into for faster pulls (GEE collection query aborts after accumulating a chunk size with over 5000 elements i.e. rows and columns combined)

gee_stat

the zonal statistic to be estimated. The current options are mean, median, min, max and stdDev

gee_dataname

character of length 1; The specific google earth engine collection dataset name

Value

an object of class data.frame