Pull spatial data from google earth engine (GEE) for a given region on any feature in the google earth engine database

Pull spatial data from google earth engine (GEE) for a given region on any feature in the google earth engine database

Usage

gee_datapull(
  email = "ifeanyi.edochie@gmail.com",
  gee_boundary = "users/ifeanyiedochie/gadm36_CMR_0",
  gee_polygons = "users/ifeanyiedochie/cmr_polypop_boundary",
  gee_dataname = "NOAA/VIIRS/DNB/MONTHLY_V1/VCMSLCFG",
  gee_datestart = "2018-01-01",
  gee_dateend = "2018-01-31",
  gee_band = "avg_rad",
  scale = 100,
  gee_desc = "nighttimelight_cmr",
  gee_stat = "mean",
  gdrive_folder = "/SAEplus",
  ldrive_dsn = "data/cmr_nighttimelight",
  gee_crs = "EPSG:4326",
  tileScale = 1
)

Arguments

email

A GEE authenticated gmail address

gee_boundary

A shapefile with the boundaries for a specific region/country for which data is to be collected

gee_polygons

A polygon shapefile spanning the region in the gee_boundary file for features will be computed. forms the unit of analysis

gee_dataname

The specific google earth engine collection dataset name (use default as example) see (https://developers.google.com/earth-engine/datasets/) for full name

gee_datestart

The starting date for the specific feature of interest

gee_dateend

The ending date for the specific feature of interest

gee_band

The bandname within the gee_dataname selected (see https://developers.google.com/earth-engine/datasets/)

scale

Used in mean region reduction for zonal statistics

gee_desc

The name to be used to name output in google drive as well as local drive

gdrive_folder

Google Drive folder name to be created or uses if already existing to store shapefile output

ldrive_dsn

Full file path (including shapefile name) for local storage of resulting shapefile

gee_crs

Set CRS

gee_image

if TRUE, an image rather than an image collection is expected (FALSE by default)

Value

shapefiles to local drive