Compute National Poverty Line

This function computes the poverty line along the order norm scale necessary for EMDI

Usage

saeplus_ordernormpl(
  npl_value = 5006362,
  pcexp = hh.dt$pcexp,
  method = c("inclusion", "interpolation", "limsup")
)

Arguments

npl_value

a numeric value for the national/international poverty line to be coverted to order norm scale for imputation

pcexp

a numeric vector of per capita expenditure values to be converted to undergo order norm normalization

method

a string value indicating the method to be used to convert the npl from currency to normalization. Three values could be selected "inclusion", "interpolation" and "limsup

Details

The "inclusion" method add the poverty line value stipulated to the vector of welfare vector provided. The order norm value obtained from the normalization process is the converted value The interpolation method uses the 'pcexp' vector as given and then uses a linear interpolation method to estimate a conversion The limsup method takes the converted value of the greatest welfare value below the poverty line